refactor factory instantiation and database update jobs
