--- evcouplings/management/database/__init__.py
@@ -1,8 +1,0 @@
-EStatus = (lambda **enums: type('Enum', (), enums))(
-    INIT="initialized",
-    PEND="pending",
-    RUN="running",
-    DONE="done",
-    FAIL="failed",  # job failed due to bug
-    TERM="terminated",  # job was terminated externally
-)
